requirements.txt
flask
flask-cors
pdfminer.six
python-docx
